Cross Validation routine
A flexible routine for cross validation any models that work with a set of parameters and output a single output (no vector yet). Also now I modified the import data to include one out of k coding for the continents which seems to be good for the regression. I've kept the old script, so we can try both.